init user framework

+++ user/src/main/java/cn/hopever/platform/user/web/controller/UserController.java
@@ -102,20 +102,26 @@ public class UserController {
 
     @PreAuthorize("#oauth2.hasScope('user_admin_client') and ( hasRole('ROLE_super_admin') or hasRole('ROLE_admin'))")
     @RequestMapping(value = "/delete", method = {RequestMethod.GET})
-    public void delete(@RequestParam Long id) {
-        this.userTableService.delete(id);
+    public void delete(@RequestParam Long id, Principal principal) {
+        if (validateUserOperation(userTableService.getUserByUsername(principal.getName()), this.userTableService.get(id))) {
+            this.userTableService.delete(id);
+        }
     }
 
     @PreAuthorize("#oauth2.hasScope('user_admin_client') and ( hasRole('ROLE_super_admin') or hasRole('ROLE_admin'))")
     @RequestMapping(value = "/info", method = {RequestMethod.GET})
-    public Map info(@RequestParam Long id) {
+    public Map info(@RequestParam Long id, Principal principal) {
         //返回user是无法解析的，要使用对象解析为map 的形式
-        UserTable ut = this.userTableService.get(id);
-        return JacksonUtil.mapper.convertValue(userTableAssembler.toResource(ut), Map.class);
+        UserTable ut = userTableService.get(id);
+        if (validateUserOperation(userTableService.getUserByUsername(principal.getName()), ut)) {
+            return JacksonUtil.mapper.convertValue(userTableAssembler.toResource(ut), Map.class);
+        } else {
+            return null;
+        }
     }
 
     @PreAuthorize("#oauth2.hasScope('user_admin_client')")
-    @RequestMapping(value = {"/",""}, method = {RequestMethod.GET})
+    @RequestMapping(value = {"/", ""}, method = {RequestMethod.GET})
     public Map info(Principal principal) {
         //返回user是无法解析的，要使用对象解析为map 的形式
         UserTable ut = this.userTableService.getUserByUsername(principal.getName());
@@ -125,31 +131,31 @@ public class UserController {
     @PreAuthorize("#oauth2.hasScope('user_admin_client')")
     @RequestMapping(value = "/personal/update", method = {RequestMethod.POST})
     public Map updatePersonalUser(@RequestBody JsonNode body, Principal principal) {
-        Map map = JacksonUtil.mapper.convertValue(body.get("data"),Map.class);
-        UserTable user=this.userTableService.get( Long.valueOf(map.get("id").toString()));
-        if(body.get("data").get("email")!=null&&!body.get("data").get("email").isNull()){
+        Map map = JacksonUtil.mapper.convertValue(body.get("data"), Map.class);
+        UserTable user = this.userTableService.get(Long.valueOf(map.get("id").toString()));
+        if (body.get("data").get("email") != null && !body.get("data").get("email").isNull()) {
             UserTable ut = this.userTableService.getUserByEmail(body.get("data").get("email").asText());
-            if(ut!=null&&!Long.valueOf(map.get("id").toString()).equals(ut.getId())){
+            if (ut != null && !Long.valueOf(map.get("id").toString()).equals(ut.getId())) {
                 Map mapReturn = new HashMap<>();
-                mapReturn.put("message","用户Email已存在");
+                mapReturn.put("message", "用户Email已存在");
                 return mapReturn;
             }
             user.setEmail(body.get("data").get("email").asText());
         }
 
-        if(body.get("data").get("phone")!=null&&!body.get("data").get("phone").isNull()){
+        if (body.get("data").get("phone") != null && !body.get("data").get("phone").isNull()) {
             UserTable ut = this.userTableService.getUserByPhone(body.get("data").get("phone").asText());
-            if(ut!=null&&!Long.valueOf(map.get("id").toString()).equals(ut.getId())){
+            if (ut != null && !Long.valueOf(map.get("id").toString()).equals(ut.getId())) {
                 Map mapReturn = new HashMap<>();
-                mapReturn.put("message","用户电话号码已存在");
+                mapReturn.put("message", "用户电话号码已存在");
                 return mapReturn;
             }
             user.setPhone(body.get("data").get("phone").asText());
         }
-        if(body.get("data").get("name")!=null&&!body.get("data").get("name").isNull()){
+        if (body.get("data").get("name") != null && !body.get("data").get("name").isNull()) {
             user.setName(body.get("data").get("name").asText());
         }
-        if(body.get("data").get("password")!=null&&!body.get("data").get("password").isNull()){
+        if (body.get("data").get("password") != null && !body.get("data").get("password").isNull()) {
             user.setPassword(passwordEncoder.encode(body.get("data").get("password").asText()));
         }
         userTableService.save(user);
@@ -160,80 +166,85 @@ public class UserController {
     @PreAuthorize("#oauth2.hasScope('user_admin_client') and ( hasRole('ROLE_super_admin') or hasRole('ROLE_admin'))")
     @RequestMapping(value = "/update", method = {RequestMethod.POST})
     public Map updateUser(@RequestBody JsonNode body, Principal principal) {
-        Map map = JacksonUtil.mapper.convertValue(body.get("data"),Map.class);
-        UserTable userController=userTableService.getUserByUsername(principal.getName());
-        UserTable user=userTableService.get( Long.valueOf(map.get("id").toString()));
-        if(body.get("data").get("email")!=null&&!body.get("data").get("email").isNull()){
+        Map map = JacksonUtil.mapper.convertValue(body.get("data"), Map.class);
+        UserTable userController = userTableService.getUserByUsername(principal.getName());
+        UserTable user = userTableService.get(Long.valueOf(map.get("id").toString()));
+        if (!validateUserOperation(userController, user)) {
+            Map mapReturn = new HashMap<>();
+            mapReturn.put("message", "无权限修改该用户");
+            return mapReturn;
+        }
+        if (body.get("data").get("email") != null && !body.get("data").get("email").isNull()) {
             UserTable ut = this.userTableService.getUserByEmail(body.get("data").get("email").asText());
-            if(ut!=null&&!Long.valueOf(map.get("id").toString()).equals(ut.getId())){
+            if (ut != null && !Long.valueOf(map.get("id").toString()).equals(ut.getId())) {
                 Map mapReturn = new HashMap<>();
-                mapReturn.put("message","用户Email已存在");
+                mapReturn.put("message", "用户Email已存在");
                 return mapReturn;
             }
             user.setEmail(body.get("data").get("email").asText());
-        }else{
+        } else {
             user.setEmail(null);
         }
 
-        if(body.get("data").get("phone")!=null&&!body.get("data").get("phone").isNull()){
+        if (body.get("data").get("phone") != null && !body.get("data").get("phone").isNull()) {
             UserTable ut = this.userTableService.getUserByPhone(body.get("data").get("phone").asText());
-            if(ut!=null&&!Long.valueOf(map.get("id").toString()).equals(ut.getId())){
+            if (ut != null && !Long.valueOf(map.get("id").toString()).equals(ut.getId())) {
                 Map mapReturn = new HashMap<>();
-                mapReturn.put("message","用户电话号码已存在");
+                mapReturn.put("message", "用户电话号码已存在");
                 return mapReturn;
             }
             user.setPhone(body.get("data").get("phone").asText());
-        }else{
+        } else {
             user.setPhone(null);
         }
-        if(body.get("data").get("authorities")!=null&&!body.get("data").get("authorities").isNull()){
+        if (body.get("data").get("authorities") != null && !body.get("data").get("authorities").isNull()) {
             List list = new ArrayList<>();
             list.add(roleTableService.get(body.get("data").get("authorities").asLong()));
             user.setAuthorities(list);
-        }else {
+        } else {
             user.setAuthorities(null);
         }
         List listPartClients = new ArrayList<>();
         listPartClients.add(clientTableService.loadClientByClientId("user_admin_client"));
-        if(user.getClients()!=null){
-            for(ClientTable ct:user.getClients()){
-                if(userController.getClients()!=null&&!userController.getClients().contains(ct)){
+        if (user.getClients() != null) {
+            for (ClientTable ct : user.getClients()) {
+                if (userController.getClients() != null && !userController.getClients().contains(ct)) {
                     listPartClients.add(ct);
                 }
             }
         }
-        if(body.get("data").get("clients")!=null&&!body.get("data").get("clients").isNull()){
-            listPartClients.addAll(clientTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("clients"),List.class)));
+        if (body.get("data").get("clients") != null && !body.get("data").get("clients").isNull()) {
+            listPartClients.addAll(clientTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("clients"), List.class)));
         }
         user.setClients(listPartClients);
 
-        List<ModuleRoleTable> listPartModileRoles =user.getModulesAuthorities();
-        if(userController.getClients()!=null&&listPartModileRoles!=null&&listPartModileRoles.size()>0){
-            for(ClientTable ct:userController.getClients()){
+        List<ModuleRoleTable> listPartModileRoles = user.getModulesAuthorities();
+        if (userController.getClients() != null && listPartModileRoles != null && listPartModileRoles.size() > 0) {
+            for (ClientTable ct : userController.getClients()) {
                 List<ModuleRoleTable> moduleRoles = ct.getModuleRoles();
-                if(moduleRoles!=null){
-                    for(ModuleRoleTable mr :moduleRoles){
+                if (moduleRoles != null) {
+                    for (ModuleRoleTable mr : moduleRoles) {
                         listPartModileRoles.remove(mr);
                     }
                 }
 
             }
         }
-        if(body.get("data").get("modulesAuthorities")!=null&&!body.get("data").get("modulesAuthorities").isNull()){
-            listPartModileRoles.addAll(moduleRoleTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("modulesAuthorities"),List.class)));
+        if (body.get("data").get("modulesAuthorities") != null && !body.get("data").get("modulesAuthorities").isNull()) {
+            listPartModileRoles.addAll(moduleRoleTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("modulesAuthorities"), List.class)));
         }
         user.setModulesAuthorities(listPartModileRoles);
-        if(body.get("data").get("limitedDate")!=null&&!body.get("data").get("limitedDate").isNull()){
+        if (body.get("data").get("limitedDate") != null && !body.get("data").get("limitedDate").isNull()) {
             user.setLimitedDate(new Date(body.get("data").get("limitedDate").asLong()));
-        }else{
+        } else {
             user.setLimitedDate(null);
         }
-        if(body.get("data").get("name")!=null&&!body.get("data").get("name").isNull()){
+        if (body.get("data").get("name") != null && !body.get("data").get("name").isNull()) {
             user.setName(body.get("data").get("name").asText());
-        }else{
+        } else {
             user.setName(null);
         }
-        if(body.get("data").get("password")!=null&&!body.get("data").get("password").isNull()){
+        if (body.get("data").get("password") != null && !body.get("data").get("password").isNull()) {
             user.setPassword(passwordEncoder.encode(body.get("data").get("password").asText()));
         }
         userTableService.save(user);
@@ -243,64 +254,64 @@ public class UserController {
     @PreAuthorize("#oauth2.hasScope('user_admin_client') and ( hasRole('ROLE_super_admin') or hasRole('ROLE_admin'))")
     @RequestMapping(value = "/save", method = {RequestMethod.POST})
     public Map saveUser(@RequestBody JsonNode body, Principal principal) {
-        Map map = JacksonUtil.mapper.convertValue(body.get("data"),Map.class);
-        UserTable user=new UserTable();
-        if(this.userTableService.getUserByUsername(body.get("data").get("username").asText())!=null){
+        Map map = JacksonUtil.mapper.convertValue(body.get("data"), Map.class);
+        UserTable user = new UserTable();
+        if (this.userTableService.getUserByUsername(body.get("data").get("username").asText()) != null) {
             Map mapReturn = new HashMap<>();
-            mapReturn.put("message","用户账号已存在");
+            mapReturn.put("message", "用户账号已存在");
             return mapReturn;
         }
-        if(body.get("data").get("email")!=null&&!body.get("data").get("email").isNull()){
-            if(this.userTableService.getUserByEmail(body.get("data").get("email").asText())!=null){
+        if (body.get("data").get("email") != null && !body.get("data").get("email").isNull()) {
+            if (this.userTableService.getUserByEmail(body.get("data").get("email").asText()) != null) {
                 Map mapReturn = new HashMap<>();
-                mapReturn.put("message","用户Email已存在");
+                mapReturn.put("message", "用户Email已存在");
                 return mapReturn;
             }
             user.setEmail(body.get("data").get("email").asText());
-        }else{
+        } else {
             user.setEmail(null);
         }
 
-        if(body.get("data").get("phone")!=null&&!body.get("data").get("phone").isNull()){
-            if(this.userTableService.getUserByPhone(body.get("data").get("phone").asText())!=null){
+        if (body.get("data").get("phone") != null && !body.get("data").get("phone").isNull()) {
+            if (this.userTableService.getUserByPhone(body.get("data").get("phone").asText()) != null) {
                 Map mapReturn = new HashMap<>();
-                mapReturn.put("message","用户电话号码已存在");
+                mapReturn.put("message", "用户电话号码已存在");
                 return mapReturn;
             }
             user.setPhone(body.get("data").get("phone").asText());
-        }else{
+        } else {
             user.setPhone(null);
         }
 
         user.setUsername(body.get("data").get("username").asText());
         user.setPassword(passwordEncoder.encode(body.get("data").get("password").asText()));
-        if(body.get("data").get("authorities")!=null&&!body.get("data").get("authorities").isNull()){
+        if (body.get("data").get("authorities") != null && !body.get("data").get("authorities").isNull()) {
             List list = new ArrayList<>();
             list.add(roleTableService.get(body.get("data").get("authorities").asLong()));
             user.setAuthorities(list);
-        }else {
+        } else {
             user.setAuthorities(null);
         }
         List clientsUpdate = new ArrayList<>();
         clientsUpdate.add(clientTableService.loadClientByClientId("user_admin_client"));
-        if(body.get("data").get("clients")!=null&&!body.get("data").get("clients").isNull()){
-            clientTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("clients"),List.class));
-            clientsUpdate.addAll(clientTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("clients"),List.class)));
+        if (body.get("data").get("clients") != null && !body.get("data").get("clients").isNull()) {
+            clientTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("clients"), List.class));
+            clientsUpdate.addAll(clientTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("clients"), List.class)));
         }
         user.setClients(clientsUpdate);
-        if(body.get("data").get("modulesAuthorities")!=null&&!body.get("data").get("modulesAuthorities").isNull()){
-            user.setModulesAuthorities(moduleRoleTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("modulesAuthorities"),List.class)));
-        }else {
+        if (body.get("data").get("modulesAuthorities") != null && !body.get("data").get("modulesAuthorities").isNull()) {
+            user.setModulesAuthorities(moduleRoleTableService.getByIds(JacksonUtil.mapper.convertValue(body.get("data").get("modulesAuthorities"), List.class)));
+        } else {
             user.setModulesAuthorities(null);
         }
-        if(body.get("data").get("limitedDate")!=null&&!body.get("data").get("limitedDate").isNull()){
+        if (body.get("data").get("limitedDate") != null && !body.get("data").get("limitedDate").isNull()) {
             user.setLimitedDate(new Date(body.get("data").get("limitedDate").asLong()));
-        }else{
+        } else {
             user.setLimitedDate(null);
         }
-        if(body.get("data").get("name")!=null&&!body.get("data").get("name").isNull()){
+        if (body.get("data").get("name") != null && !body.get("data").get("name").isNull()) {
             user.setName(body.get("data").get("name").asText());
-        }else{
+        } else {
             user.setName(null);
         }
         user.setCreatedDate(new Date());
@@ -309,16 +320,25 @@ public class UserController {
         return null;
     }
 
-    private static boolean validateUserOperation(UserTable ut1,UserTable ut2){
-        if(ut1.getAuthorities().get(0).getAuthority().equals("ROLE_super_admin")){
+    private static boolean validateUserOperation(UserTable ut1, UserTable ut2) {
+        if (ut1.getAuthorities().get(0).getAuthority().equals("ROLE_super_admin")) {
             return true;
-        }else{
-            if(ut1.getAuthorities().get(0).getAuthority().equals("ROLE_admin")&&ut2.getAuthorities().get(0).getAuthority().equals("ROLE_admin")){
-                if(ut2.getCreateUser()!=null&&ut2.getCreateUser().getId() ==ut1.getId()){
+        } else {
+            if (ut1.getAuthorities().get(0).getAuthority().equals("ROLE_admin") && ut2.getAuthorities().get(0).getAuthority().equals("ROLE_admin")) {
+                if (ut2.getCreateUser() != null && ut2.getCreateUser().getId() == ut1.getId()) {
                     return true;
                 }
-            }else if(ut1.getAuthorities().get(0).getAuthority().equals("ROLE_admin")&&ut2.getAuthorities().get(0).getAuthority().equals("ROLE_common_user")){
-                //for()
+            } else if (ut1.getAuthorities().get(0).getAuthority().equals("ROLE_admin") && ut2.getAuthorities().get(0).getAuthority().equals("ROLE_common_user")) {
+                if (ut1.getClients() != null) {
+                    for (ClientTable ct : ut1.getClients()) {
+                        if(ct.getClientId().equals("user_admin_client")){
+                            continue;
+                        }
+                        if (ut2.getClients() != null && ut2.getClients().contains(ct)) {
+                            return true;
+                        }
+                    }
+                }
             }
             return false;
         }
